requirements.txt
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
Flask-Migrate==2.1.1
Flask-WTF==0.14.2
psycopg2-binary==2.7.4
Flask-Login==0.4.1
Flask-CLI==0.4.0
requests_oauthlib
google-cloud-bigquery
